graph-datasources/hybrid_search/requirements.txt
llama-index-core
llama-index-utils-workflow
llama-index-vector-stores-milvus
llama-index-graph-stores-memgraph
python-dotenv
fastapi
uvicorn[standard]
pydantic
pymilvus
redis
requests


